requirements.txt
# TODO: Specify deps necessary
celery>5
daphne==4.0.0
Django>4.1
gunicorn>20
psycopg2==2.9.5
redis~=4.3
environ~=1.0
requests~=2.28.1
djangorestframework~=3.14.0
asgiref~=3.5.2